Refactor e2e test utilities and fix service window restore
- Extract gotoAndSettle + waitForStateChange into utils/pageHelper.ts
- Add utils/fixtures.ts with ordersOpen fixture (replaces try/finally pattern)
- Fix service window restore bug: selector now targets standalone OrdersOpenSwitch when orders are open, not the closed-state alert
- Fix global-teardown storageFile path to resolve relative to file location
- Auto-create e2e/data/ directory in orderHelper to avoid crash on fresh clone
- Expand .gitignore to cover qa/prod storage files and playwright output dirs
